scratch-messaging: fix profile comments with linebreaks (#3341)
* comments-linebreaks: split files

* scratch-messaging: fix profile comments with linebreaks

* Format code
