+++ addons/comments-linebreaks/break-spaces.css
@@ -1,0 +1,4 @@
+.comment-content,
+.comment .content {
+  white-space: break-spaces;
+}
